src/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1 #
djangorestframework==3.14.0 #
djangorestframework-simplejwt==5.3.1 #
pillow==10.4.0
drf-yasg==1.21.7

requests~=2.32.3
